Conflicts:
	include/diane_octomap/diane_octomap_nodelet.h
	src/lib/diane_octomap_nodelet.cpp

+++ src/lib/diane_octomap.cpp
@@ -455,7 +455,7 @@ vector<diane_octomap::Line*> diane_octomap::DianeOctomap::CreateGroupLines(vecto
 {
     vector<Line*> Lines;
 
-    double Min_Votes = 10;
+    int Min_Votes = 10;
 
     for (int i=0; i<Votes.size(); i++)
     {
@@ -472,6 +472,7 @@ vector<diane_octomap::Line*> diane_octomap::DianeOctomap::CreateGroupLines(vecto
 
                 Lines.push_back(line);
             }
+
 
         }
 
@@ -1329,6 +1330,8 @@ void diane_octomap::DianeOctomap::WriteStairCandidateToFileWithMatrix(diane_octo
     stairfile.close();
 
 }
+
+
 
 
 //-------------------------------------------------------------------------------------------------------------------------------------
